+++ src/retry/src/backoff_policy.rs
@@ -63,7 +63,7 @@ pub trait BackoffPolicy: Send + Sync + std::fmt::Debug {
 
 /// A helper type to use [BackoffPolicy] in client and request options.
 #[derive(Clone)]
-pub struct BackoffPolicyArg(pub(crate) Arc<dyn BackoffPolicy>);
+pub struct BackoffPolicyArg(pub Arc<dyn BackoffPolicy>);
 
 impl<T: BackoffPolicy + 'static> std::convert::From<T> for BackoffPolicyArg {
     fn from(value: T) -> Self {
